refactor: move migration JNI exports into crate::jni
Move `MigrationRustBackend`'s 32 exports into `jni/migration.rs`, leaving
`crate::migration` as the engine, planning, proving and wallet logic they
call into. That module drops from 3214 to 1423 lines and no longer
references the `jni` crate outside its tests, so its live-wallet test
suite exercises ordinary Rust.

The Java class descriptors and object encoders move with the exports.
They outgrew a single file, so they sit in `jni/migration/encode.rs`,
scoped `pub(super)` and therefore reachable only from the migration
exports rather than from every future submodule of `crate::jni`.

`decode_byte_array_list` is the exception: it decodes a `JObjectArray` of
byte arrays with no migration-specific behaviour, so it joins the other
domain-neutral marshalling helpers in `crate::utils`.

Twelve items in `crate::migration` widen from private to `pub(crate)` to
stay reachable from the exports. Each has a call site; none is dead.

Exported symbol names are unchanged, verified by diffing `nm -gU` output
against a pre-refactor baseline of 187 `Java_*` symbols.
